Updating warning message instead of rebuilding dom
Instead of destroying the warning message dom elements every 4 seconds this updates just the warning message. 

First the warning is hidden. Then it is checked if the warning should display. If it should be displayed we show the warning. We then update the warning message string.

This stops the warning from blinking when the user is hovering it.